# backend/tests/contract/snapshots/api-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "CreateExportRequest": {
        "additionalProperties": false,
        "properties": {
          "reviewed_result_id": {
            "format": "uuid",
            "title": "Reviewed Result Id",
            "type": "string"
          }
        },
        "required": [
          "reviewed_result_id"
        ],
        "title": "CreateExportRequest",
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "blockers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blockers"
          },
          "cause_category": {
            "title": "Cause Category",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "location_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Ref"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase"
          },
          "project_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "retryable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retryable"
          },
          "severity": {
            "enum": [
              "fatal",
              "blocking",
              "review_required",
              "warning",
              "informational"
            ],
            "title": "Severity",
            "type": "string"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "workbench_ready": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workbench Ready"
          }
        },
        "required": [
          "code",
          "message",
          "severity",
          "stage",
          "location_ref",
          "cause_category"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ExportArtifactResponse": {
        "properties": {
          "downloadable": {
            "title": "Downloadable",
            "type": "boolean"
          },
          "kind": {
            "enum": [
              "ballooned_pdf",
              "sip_excel",
              "manifest"
            ],
            "title": "Kind",
            "type": "string"
          },
          "reviewed_result_id": {
            "format": "uuid",
            "title": "Reviewed Result Id",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "kind",
          "sha256",
          "size_bytes",
          "reviewed_result_id",
          "downloadable"
        ],
        "title": "ExportArtifactResponse",
        "type": "object"
      },
      "ExportResponse": {
        "properties": {
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ExportArtifactResponse"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "error_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mapping_version": {
            "title": "Mapping Version",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "renderer_version": {
            "title": "Renderer Version",
            "type": "string"
          },
          "reviewed_result_id": {
            "format": "uuid",
            "title": "Reviewed Result Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "running",
              "success",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "template_version": {
            "title": "Template Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "reviewed_result_id",
          "status",
          "error_id",
          "template_version",
          "mapping_version",
          "renderer_version",
          "artifacts"
        ],
        "title": "ExportResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Quality Inspection",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/exports/{export_id}": {
      "get": {
        "operationId": "QI-API-EXP-002",
        "parameters": [
          {
            "in": "path",
            "name": "export_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Export Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Requested resource was not found.",
            "x-stable-error-codes": [
              "project_not_found",
              "export_not_found"
            ]
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request failed transport or business validation.",
            "x-stable-error-codes": [
              "request_validation_failed"
            ]
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected internal failure.",
            "x-stable-error-codes": [
              "internal_server_error"
            ]
          }
        },
        "summary": "Get Export",
        "tags": [
          "exports"
        ]
      }
    },
    "/api/v1/exports/{export_id}/downloads/{kind}": {
      "get": {
        "operationId": "QI-API-EXP-003",
        "parameters": [
          {
            "in": "path",
            "name": "export_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Export Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "application/pdf": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Published export artifact."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Requested resource was not found.",
            "x-stable-error-codes": [
              "project_not_found",
              "export_artifact_not_found"
            ]
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request conflicts with current aggregate or capability state.",
            "x-stable-error-codes": [
              "export_artifact_unavailable"
            ]
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request failed transport or business validation.",
            "x-stable-error-codes": [
              "request_validation_failed"
            ]
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected internal failure.",
            "x-stable-error-codes": [
              "internal_server_error"
            ]
          }
        },
        "summary": "Download Export",
        "tags": [
          "exports"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/exports": {
      "post": {
        "operationId": "QI-API-EXP-001",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Requested resource was not found.",
            "x-stable-error-codes": [
              "project_not_found",
              "reviewed_result_not_found"
            ]
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request conflicts with current aggregate or capability state.",
            "x-stable-error-codes": [
              "export_in_progress",
              "export_preflight_failed",
              "export_template_unavailable",
              "export_template_mapping_unavailable",
              "export_font_unavailable",
              "export_font_license_unavailable",
              "export_template_mapping_hash_mismatch",
              "export_template_hash_mismatch",
              "export_template_registration_invalid",
              "export_template_invalid",
              "export_template_sheet_missing",
              "export_font_hash_mismatch",
              "export_font_license_hash_mismatch"
            ]
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request failed transport or business validation.",
            "x-stable-error-codes": [
              "request_validation_failed"
            ]
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected internal failure.",
            "x-stable-error-codes": [
              "internal_server_error"
            ]
          }
        },
        "summary": "Create Export",
        "tags": [
          "exports"
        ]
      }
    },
...
  }
}